# Improve eval check performance by precomputing powers of poly_mix