--- a/risc0/circuit/recursion/src/cpu.rs
+++ b/risc0/circuit/recursion/src/cpu.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 RISC Zero, Inc.
+// Copyright 2024 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@ use risc0_zkp::{
     core::log2_ceil,
     field::{
         baby_bear::{BabyBear, BabyBearElem, BabyBearExtElem},
-        Elem, ExtElem, RootsOfUnity,
+        map_pow, Elem, ExtElem, RootsOfUnity,
     },
     hal::{cpu::CpuBuffer, CircuitHal, Hal},
     INV_RATE,
@@ -59,6 +59,8 @@ where
         const EXP_PO2: usize = log2_ceil(INV_RATE);
         let domain = steps * INV_RATE;
 
+        let poly_mix_pows = map_pow(poly_mix, crate::info::POLY_MIX_POWERS);
+
         // SAFETY: Convert a borrow of a cell into a raw const slice so that we can pass
         // it over the thread boundary. This should be safe because the scope of the
         // usage is within this function and each thread access will not overlap with
@@ -76,11 +78,12 @@ where
         let out = unsafe { std::slice::from_raw_parts(out.as_ptr(), out.len()) };
         let check = check.as_slice();
         let check = unsafe { std::slice::from_raw_parts(check.as_ptr(), check.len()) };
+        let poly_mix_pows = poly_mix_pows.as_slice();
 
         let args: &[&[BabyBearElem]] = &[&code, &out, &data, &mix, &accum];
 
         (0..domain).into_par_iter().for_each(|cycle| {
-            let tot = self.circuit.poly_fp(cycle, domain, &poly_mix, args);
+            let tot = self.circuit.poly_fp(cycle, domain, poly_mix_pows, args);
             let x = BabyBearElem::ROU_FWD[po2 + EXP_PO2].pow(cycle);
             // TODO: what is this magic number 3?
             let y = (BabyBearElem::new(3) * x).pow(1 << po2);