--- a/risc0/circuit/recursion/src/metal.rs
+++ b/risc0/circuit/recursion/src/metal.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 RISC Zero, Inc.
+// Copyright 2024 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ use risc0_zkp::{
     core::log2_ceil,
     field::{
         baby_bear::{BabyBearElem, BabyBearExtElem},
-        RootsOfUnity,
+        map_pow, RootsOfUnity,
     },
     hal::{
         metal::{BufferImpl as MetalBuffer, MetalHal, MetalHash},
@@ -63,10 +63,14 @@ impl<MH: MetalHash> CircuitHal<MetalHal<MH>> for MetalCircuitHal<MH> {
     ) {
         const EXP_PO2: usize = log2_ceil(INV_RATE);
         let domain = steps * INV_RATE;
-        let poly_mix =
-            MetalBuffer::copy_from(&self.hal.device, self.hal.cmd_queue.clone(), &[poly_mix]);
         let rou = BabyBearElem::ROU_FWD[po2 + EXP_PO2];
         let rou = MetalBuffer::copy_from(&self.hal.device, self.hal.cmd_queue.clone(), &[rou]);
+        let poly_mix_pows = map_pow(poly_mix, crate::info::POLY_MIX_POWERS);
+        let poly_mix_pows = MetalBuffer::copy_from(
+            &self.hal.device,
+            self.hal.cmd_queue.clone(),
+            poly_mix_pows.as_slice(),
+        );
         let po2 =
             MetalBuffer::copy_from(&self.hal.device, self.hal.cmd_queue.clone(), &[po2 as u32]);
         let size = MetalBuffer::copy_from(
@@ -81,7 +85,7 @@ impl<MH: MetalHash> CircuitHal<MetalHal<MH>> for MetalCircuitHal<MH> {
             groups[REGISTER_GROUP_ACCUM].as_arg(),
             globals[GLOBAL_MIX].as_arg(),
             globals[GLOBAL_OUT].as_arg(),
-            poly_mix.as_arg(),
+            poly_mix_pows.as_arg(),
             rou.as_arg(),
             po2.as_arg(),
             size.as_arg(),