--- a/risc0/circuit/rv32im-sys/cxx/step_verify_accum.cpp
+++ b/risc0/circuit/rv32im-sys/cxx/step_verify_accum.cpp
@@ -56,10 +56,10 @@ Fp step_verify_accum(void* ctx, HostBridge host, size_t steps, size_t cycle, Fp*
   // loc("Top/Code/OneHot/hot[6](Reg)"("./zirgen/components/mux.h":49:25))
   auto x8 = args[0][7 * steps + ((cycle - 0) & mask)];
   assert(x8 != Fp::invalid());
-  // loc("Top/Code/OneHot/hot[0](Reg)"("zirgen/circuit/rv32im/top.cpp":78:27))
+  // loc("Top/Code/OneHot/hot[0](Reg)"("zirgen/circuit/rv32im/v1/edsl/top.cpp":78:27))
   auto x9 = args[0][1 * steps + ((cycle - 0) & mask)];
   assert(x9 != Fp::invalid());
-  // loc("Top/Code/OneHot/hot[2](Reg)"("zirgen/circuit/rv32im/top.cpp":78:27))
+  // loc("Top/Code/OneHot/hot[2](Reg)"("zirgen/circuit/rv32im/v1/edsl/top.cpp":78:27))
   auto x10 = args[0][3 * steps + ((cycle - 0) & mask)];
   assert(x10 != Fp::invalid());
   if (x9 != 0) {