--- a/risc0/circuit/rv32im-sys/src/ffi.rs
+++ b/risc0/circuit/rv32im-sys/src/ffi.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 RISC Zero, Inc.
+// Copyright 2024 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.