--- a/risc0/circuit/rv32im/src/cuda.rs
+++ b/risc0/circuit/rv32im/src/cuda.rs
@@ -17,7 +17,7 @@ use std::rc::Rc;
 use cust::prelude::*;
 use risc0_core::field::{
     baby_bear::{BabyBearElem, BabyBearExtElem},
-    RootsOfUnity,
+    map_pow, Elem, ExtElem, RootsOfUnity,
 };
 use risc0_zkp::{
     core::log2_ceil,
@@ -81,10 +81,22 @@ impl<'a, CH: CudaHash> CircuitHal<CudaHal<CH>> for CudaCircuitHal<CH> {
         let domain = steps * INV_RATE;
         let rou = BabyBearElem::ROU_FWD[po2 + EXP_PO2];
 
-        let poly_mix = self.hal.copy_from_extelem("poly_mix", &[poly_mix]);
         let rou = self.hal.copy_from_elem("rou", &[rou]);
         let po2 = self.hal.copy_from_u32("po2", &[po2 as u32]);
         let size = self.hal.copy_from_u32("size", &[domain as u32]);
+
+        let poly_mix_pows = map_pow(poly_mix, crate::info::POLY_MIX_POWERS);
+        let poly_mix_pows: &[u32; BabyBearExtElem::EXT_SIZE * crate::info::NUM_POLY_MIX_POWERS] =
+            BabyBearExtElem::as_u32_slice(poly_mix_pows.as_slice())
+                .try_into()
+                .unwrap();
+
+        let mix_pows_name = std::ffi::CString::new("poly_mix").unwrap();
+        self.module
+            .get_global(&mix_pows_name)
+            .unwrap()
+            .copy_from(poly_mix_pows)
+            .unwrap();
 
         let stream = Stream::new(StreamFlags::DEFAULT, None).unwrap();
 
@@ -98,7 +110,6 @@ impl<'a, CH: CudaHash> CircuitHal<CudaHal<CH>> for CudaCircuitHal<CH> {
                 accum.as_device_ptr(),
                 mix.as_device_ptr(),
                 out.as_device_ptr(),
-                poly_mix.as_device_ptr(),
                 rou.as_device_ptr(),
                 po2.as_device_ptr(),
                 size.as_device_ptr()