… risc0/circuit/recursion-sys/cxx/poly_fp.cpp
… risc0/circuit/recursion-sys/cxx/step_exec.cpp
… risc0/circuit/recursion-sys/kernels/cuda/eval_check.cu
… risc0/circuit/recursion-sys/kernels/metal/eval_check.metal
… risc0/circuit/recursion/src/info.rs
… risc0/circuit/recursion/src/recursion_zkr.zip
… risc0/circuit/rv32im-sys/cxx/poly_fp.cpp
… risc0/circuit/rv32im-sys/cxx/step_exec.cpp
… risc0/circuit/rv32im-sys/kernels/cuda/eval_check.cu
… risc0/circuit/rv32im-sys/kernels/metal/eval_check.metal
… risc0/circuit/rv32im/src/info.rs
… risc0/circuit/rv32im/src/poly_ext.rs