CUDA:
```
fib/100/prove
-------------
base        1.00     672.7±3.76ms  95.1 KElem/sec
changes     1.65    1107.9±3.14ms  57.8 KElem/sec

fib/1000/prove
--------------
base        1.00     703.6±3.33ms  91.0 KElem/sec
changes     1.61    1133.8±4.53ms  56.4 KElem/sec

fib/10000/prove
---------------
changes     1.00        2.1±0.01s  120.9 KElem/sec
base        1.03        2.2±0.01s  117.3 KElem/sec
```
CPU:
```
fib/100/prove
-------------
changes     1.00        2.9±0.01s  22.4 KElem/sec
base        1.46        4.2±0.01s  15.3 KElem/sec

fib/1000/prove
--------------
changes     1.00        2.9±0.01s  22.1 KElem/sec
base        1.46        4.2±0.01s  15.2 KElem/sec

fib/10000/prove
---------------
changes     1.00       11.6±0.03s  22.2 KElem/sec
base        1.47       17.0±0.03s  15.1 KElem/sec

```