@shkoo 2024-03-11
See also risc0/risczero-wip#512