@flaub 2024-03-12
This seems to include the circom fixes, I wonder if we should land that first, independently (cc: @nategraf)