@shkoo 2024-03-12
Reverted the circom change